Add annotation to allow spreading VIPs across keepalived instances
A common use case for keepalived in environments with higher load is to
have multiple VIPs for HA configured in such a way that they are spread
across the nodes. This configuration allows for maximum load balancing, for
example through DNS round robin, while maintaining HA, as a failure of a
node will still result in another node getting ownership of the VIP.

This commit introduces the possibility to specify an extra annotation
on services requesting keepalived-managed IPs, that will instruct
keepalived-operator to spread the VIPs of the service across all
of the keepalived instances of the KeepalivedGroup. If the annotation
is found, keepalived-operator will add extra configs to set a preferred
owner for each VIP in the service, using a round-robin algorithm among
the keepalived pods. The commit also adds the ability to watch for pod
changes, to ensure that any change in the pods backing the daemonset
is reflected in the configurations.

This commit aims at being 100% backwards-compatible, by adding this
functionality as an opt-in option. Services that do not specify the
new annotation will continue to behave as they always have.