Handle local externalTrafficPolicy correctly
LoadBalancer services can specify an extra property called `externalTrafficPolicy`
which defines how kube-proxy sets up the load balancing on the nodes. If such
property is set to `Local`, kube-proxy will only forward traffic for externalIPs
or LoadBalancer IPs to pods backing the service that are running on the node on
which the packet arrived. If no pods are running on that node, that traffic gets
dropped.

This commit introduces a vrrp_script for services that set this property to
`Local` which will healthcheck the Kubernetes provided health endpoint for the
service. This endpoint returns 200 if at least a pod backing the service is
running on the node, or 503 if no pods are running on the node. By adding this
healthcheck script, keepalived will deem the node unhealthy if the healthcheck
fails, thus ensuring the VIP moves to a different owner that does have at least
a pod backing the service running on it. This way no traffic will be lost and
the `Local` `externalTrafficPolicy` will work as intended.