Enable interface autodiscovery based on a reachable IP
Keepalived-operator assumes that all nodes in a given KeepalivedGroup
will have the same networking configurations and the same interface names.
This is not always true and it makes for a harder setup in non-omogeneus
environments.

This commit introduces an extra optional field on the CRD called
`interfaceFromIP`. If this field is specified, the `interface` field will
be ignored and the interface name will instead be autodiscovered on each
node based on which interface would be used to reach the specified IP.

This commit requires the keepalived-operator image to contain the `ip`, `sed`
and `awk` binaries on the path.